+++ src/Lakona.Hub/HubJsonContext.cs
@@ -1,4 +1,5 @@
 using System.Text.Json.Serialization;
+using Lakona.Hub.Applications;
 using Lakona.Hub.Updates;
 
 namespace Lakona.Hub;
@@ -7,5 +8,6 @@ namespace Lakona.Hub;
     PropertyNamingPolicy = JsonKnownNamingPolicy.CamelCase,
     WriteIndented = true)]
 [JsonSerializable(typeof(Dictionary<string, string>), TypeInfoPropertyName = "ApplicationPaths")]
+[JsonSerializable(typeof(StoredManualApplicationSettings))]
 [JsonSerializable(typeof(HubReleaseManifest))]
 internal sealed partial class HubJsonContext : JsonSerializerContext;